# Decrease xlsx reader memory consumption