--- a/src/PhpSpreadsheet/Reader/Xlsx/SheetStructure.php
+++ b/src/PhpSpreadsheet/Reader/Xlsx/SheetStructure.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheet\Reader\Xlsx;
+
+class SheetStructure
+{
+    public const CONDITIONAL_FORMATTING = 'conditionalFormatting';
+    public const SHEET_PROTECTION = 'sheetProtection';
+    public const SHEET_VIEW = 'sheetView';
+    public const MERGE_CELL = 'mergeCell';
+    public const EXT_LST = 'extLst';
+    public const DATA_VALIDATION = 'dataValidation';
+    public const LEGACY_DRAWING_HF = 'legacyDrawingHF';
+    public const DRAWING = 'drawing';
+    public const AUTO_FILTER = 'autoFilter';
+    public const TABLE_PART = 'tablePart';
+    public const HYPERLINK = 'hyperlink';
+    public const ROW = 'row';
+    public const COL = 'col';
+    public const PROTECTED_RANGE = 'protectedRange';
+    public const SHEET_PR = 'sheetPr';
+    public const SHEET_FORMAT_PR = 'sheetFormatPr';
+    public const PRINT_OPTIONS = 'printOptions';
+    public const PAGE_MARGINS = 'pageMargins';
+    public const PAGE_SETUP = 'pageSetup';
+    public const HEADER_FOOTER = 'headerFooter';
+    public const ROW_BREAKS = 'rowBreaks';
+    public const COL_BREAKS = 'colBreaks';
+    public const ALTERNATE_CONTENT = 'AlternateContent';
+}