This is:

```
- [ ] a bugfix
- [ ] a new feature
- [x] refactoring
```

Checklist:

- [x] Changes are covered by unit tests
- [x] Code style is respected
- [x] Commit message explains **why** the change is made (see https://github.com/erlang/otp/wiki/Writing-good-commit-messages)
- [x] CHANGELOG.md contains a short summary of the change
- [ ] Documentation is updated as necessary

### Why this change is needed?
While reading big xlsx file like 400k rows [test.xlsx](https://github.com/PHPOffice/PhpSpreadsheet/files/8464853/test.xlsx) `simplexml_load_string()` uses almost 6 GB of memory.
After changing it to `XMLReader` each node is reading separately and xml string is added to "map". Each map's element is parsed with `simplexml_load_string()` separately which dramatically decreases memory usage (~1,2 GB for the mentioned earlier file).

### How to check

XlsxTest.php
```php
<?php
require 'vendor/autoload.php';

use PhpOffice\PhpSpreadsheet\IOFactory;

$file = './test.xlsx';

$inputFileType = IOFactory::identify($file);
$objReader = IOFactory::createReader($inputFileType);
$excel = $objReader->load($file);

echo 'Done';
```

```
/usr/bin/time -l php XlsxTest.php
```
(cannot use `memory_get_usage()` because of https://bugs.php.net/bug.php?id=62467)