@MarkBaker 2022-05-01
Sorry that it's taken me a while to get round to looking at this: I've been doing a lot of work elsewhere in the application, most recently in reducing memory usage for the cell collection once a file is loaded.

I'm about to take a deeper dive into this PR (thank you for the submission), and it certainly looks positive.

I can't even load your sample test file with my basic test setup;
```
vagrant@homestead:~/projects/PHPOffice/PHPSpreadsheet/develop$ php samples/readerTest.php
Killed
```
But using your branch, the file can actually be loaded
```
vagrant@homestead:~/projects/PHPOffice/PHPSpreadsheet/develop$ php samples/readerTest.php

Call time to load spreadsheet file was 188.0361 seconds
 Current memory usage: 1064964 KB
    Peak memory usage: 1087496 KB
```
and that's a very powerful positive.

I'll need to do some more testing with smaller files, and with a range of features, so that I can make a proper comparison of speed and memory usage; and also take a look at how you've made the changes.