@MarkBaker 2022-05-01
Just out of interest for a comparison: if I merge the memory-saving changes that I've already merged to master into your branch, I get the following memory usage:
```
vagrant@homestead:~/projects/PHPOffice/PHPSpreadsheet/develop$ php samples/readerTest.php

Call time to load spreadsheet file was 197.1121 seconds
 Current memory usage: 1001476 KB
    Peak memory usage: 1024008 KB
```
Reduced by another 60MB