@MarkBaker 2022-05-01
There are definitely some trade-offs here:

Some tests with smaller files, so that I can do an easy comparison between master and your code. With a 10-column, 40k row spreadsheet with purely numeric cell values:

master
```
Call time to load spreadsheet file was 30.4329 seconds
 Current memory usage: 147456 KB
    Peak memory usage: 147456 KB
```
your PR
```
Call time to load spreadsheet file was 38.3924 seconds
 Current memory usage: 186368 KB
    Peak memory usage: 188420 KB
```

With a 10-column, 102k row spreadsheet with purely numeric cell values:

master
```
Call time to load spreadsheet file was 96.7067 seconds
 Current memory usage: 333824 KB
    Peak memory usage: 333824 KB
```
your PR
```
Call time to load spreadsheet file was 95.7461 seconds
 Current memory usage: 434176 KB
    Peak memory usage: 438276 KB
```

But if I increase that to 200k rows, then master crashes out but your PR will successfully read the file
```
Call time to load spreadsheet file was 208.5629 seconds
 Current memory usage: 864256 KB
    Peak memory usage: 872452 KB
```

Your approach performs about 25% slower with smaller spreadsheets until about the million cell mark; and is consistently higher memory consumption where I'm able to measure both; yet it's able to read the largest spreadsheets where the current master branch code fails.

That gives me something of a dilemma. I'd love to be able to handle those extremely large spreadsheets as well; but not at the expense of higher memory usage and slower performance for the majority of users working with smaller spreadsheets. Your PR is great for the exception cases; but not for the majority of cases.

Having two Xlsx Readers, one for the smaller files and one for extremely large files might be an option, although it would have a lot of code duplication between the two, and increase the maintenance work; but it's also difficult to advise the circumstances when each should be used... it's not based purely on number of cells, shared strings and shared formulae make a difference.